add ft_strchr

+++ libft.h
@@ -2,8 +2,10 @@
 #define LIBTF_H
 
 #include <stddef.h>
+#include <stdio.h>
+#include <stdio.h>
 
-int	ft_strlen(char *c);
+size_t	ft_strlen(char *c);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 size_t	strlcat(char *dst, const char *src, size_t size);
 int	ft_toupper(int ch);
